deep_search/requirements.txt
django==1.10.2
djangorestframework==3.5.4
pillow==4.0.0